Long uses now INT type

+++ src/main/java/com/p000ison/dev/sqlapi/mysql/MySQLTableBuilder.java
@@ -46,48 +46,52 @@ public final class MySQLTableBuilder extends TableBuilder {
         StringBuilder query = new StringBuilder();
         query.append(column.getName()).append(' ');
 
-        boolean allowModifyLength = true;
+        if (column.isID()) {
+              query.append("INTEGER");
+        } else {
+            boolean allowModifyLength = true;
 
-        if (type == boolean.class || type == Boolean.class) {
-            query.append("TINYINT(1)");
-            allowModifyLength = false;
-        } else if (type == byte.class || type == Byte.class) {
-            query.append("TINYINT");
-            allowModifyLength = false;
-        } else if (type == short.class || type == Short.class) {
-            query.append("SMALLINT");
-            allowModifyLength = false;
-        } else if (type == int.class || type == Integer.class) {
-            query.append("INTEGER");
-        } else if (type == float.class || type == Float.class) {
-            query.append("FLOAT");
-        } else if (type == double.class || type == Double.class) {
-            query.append("DOUBLE");
-        } else if (type == long.class || type == Long.class) {
-            query.append("LONG");
-        } else if (type == char.class || type == Character.class) {
-            query.append("CHAR");
-        } else if (type == Date.class || type == Timestamp.class) {
-            query.append("DATETIME");
-        } else if (type == String.class) {
-            if (column.getLength().length != 0) {
-                query.append("VARCHAR");
-            } else {
-                query.append("TEXT");
-            }
-        } else if (RegisteredTable.isSerializable(type)) {
-            query.append("BLOB");
-            allowModifyLength = false;
-        }
-
-        if (column.getLength().length != 0 && allowModifyLength) {
-            query.append('(');
-            for (int singleLength : column.getLength()) {
-                query.append(singleLength).append(',');
+            if (type == boolean.class || type == Boolean.class) {
+                query.append("TINYINT(1)");
+                allowModifyLength = false;
+            } else if (type == byte.class || type == Byte.class) {
+                query.append("TINYINT");
+                allowModifyLength = false;
+            } else if (type == short.class || type == Short.class) {
+                query.append("SMALLINT");
+                allowModifyLength = false;
+            } else if (type == int.class || type == Integer.class || type == long.class || type == Long.class) {
+                query.append("INTEGER");
+            } else if (type == float.class || type == Float.class) {
+                query.append("FLOAT");
+            } else if (type == double.class || type == Double.class) {
+                query.append("DOUBLE");
+            } else if (type == char.class || type == Character.class) {
+                query.append("CHAR");
+            } else if (type == Date.class || type == Timestamp.class) {
+                query.append("DATETIME");
+            } else if (type == String.class) {
+                if (column.getLength().length != 0) {
+                    query.append("VARCHAR");
+                } else {
+                    query.append("TEXT");
+                }
+            } else if (RegisteredTable.isSerializable(type)) {
+                System.out.println(column);
+                query.append("BLOB");
+                allowModifyLength = false;
             }
 
-            query.deleteCharAt(query.length() - 1);
-            query.append(')');
+
+            if (column.getLength().length != 0 && allowModifyLength) {
+                query.append('(');
+                for (int singleLength : column.getLength()) {
+                    query.append(singleLength).append(',');
+                }
+
+                query.deleteCharAt(query.length() - 1);
+                query.append(')');
+            }
         }
 
         if (column.isNotNull()) {
